move leave notification to applier to doc_event
